pyproject.toml
[project]
name = "zotero-google-scholar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "William C Grisaitis", email = "wgrisaitis@gmail.com" }
]
requires-python = ">=3.12.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "python-dotenv>=1.1.1",
    "pyzotero>=1.6.11",
    "requests>=2.32.4",
    "scholarly>=1.7.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "detect-secrets==1.5.0",
    "pre-commit>=4.2.0",
    "ruff>=0.12.0",
]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
]
typing = [
    "mypy>=1.16.1",
]

[tool.mypy]
install_types = true
non_interactive = true

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"

[project.scripts]
zotero-google-scholar = "zotero_google_scholar.cli:main"
